+++ lib/zakuro/calculation/monthly/operated_month.rb
@@ -1,7 +1,8 @@
 # frozen_string_literal: true
 
-require_relative './month'
 require_relative '../../operation/operation'
+require_relative './operated_solar_term'
+require_relative './month'
 
 # :nodoc:
 module Zakuro
@@ -83,7 +84,7 @@ module Zakuro
 
           return unless matched
 
-          @solar_terms = self.class.create_operated_solar_term(
+          @solar_terms = OperatedSolarTerm.create_operated_solar_term(
             operated_solar_term: solar_term,
             solar_terms: solar_terms
           )
@@ -172,6 +173,15 @@ module Zakuro
         end
 
         class << self
+          #
+          # 計算結果を該当のフィールドの運用情報に書き換える
+          #
+          # @param [Operation::Diffs] month_diff 月差分
+          # @param [MonthLabel] month_label 月表示情報
+          # @param [String] name フィールド名
+          #
+          # @return [Object] 運用情報
+          #
           def rewrite_month_fields(month_diff:, month_label:, name:)
             method_name = name.intern
 
@@ -180,67 +190,6 @@ module Zakuro
             return default if diff.invalid?
 
             diff.dest
-          end
-
-          #
-          # 運用情報によって改変した二十四節気を作成する
-          #
-          # @param [<Type>] operated_solar_term 運用情報（二十四節気）
-          # @param [Array<SolarTerm>] solar_terms 二十四節気
-          #
-          # @return [Array<SolarTerm>] 二十四節気
-          #
-          def create_operated_solar_term(operated_solar_term:, solar_terms:)
-            index = operated_solar_term.index
-
-            # 別の月に移動する二十四節気。割り当てない。
-            result = remove_solar_term(index: index, solar_terms: solar_terms)
-
-            return result if used_solar_term?(index: index, solar_terms: solar_terms)
-
-            result.push(operated_solar_term)
-
-            result
-          end
-
-          # :reek:ControlParameter
-
-          #
-          # 二十四節気が使用されているか
-          #
-          # @param [Integer] index 二十四節気番号
-          # @param [Array<SolarTerm>] solar_terms 二十四節気
-          #
-          # @return [True] 使用
-          # @return [False] 未使用
-          #
-          def used_solar_term?(index:, solar_terms:)
-            solar_terms.each do |solar_term|
-              return true if index == solar_term.index
-            end
-
-            false
-          end
-
-          # :reek:ControlParameter
-
-          #
-          # 対象の二十四節気を除外する
-          #
-          # @param [Integer] index 二十四節気番号
-          # @param [Array<SolarTerm>] solar_terms 二十四節気
-          #
-          # @return [Array<SolarTerm>] 二十四節気（対象除外済）
-          #
-          def remove_solar_term(index:, solar_terms:)
-            result = []
-            solar_terms.each do |solar_term|
-              next if index == solar_term.index
-
-              result.push(solar_term)
-            end
-
-            result
           end
         end
